overlay hightlight on file-pick
