+++ src/popup/search/pages/search/Search.vue
@@ -18,7 +18,6 @@
           v-model:query="internalQuery"
           style="grid-area: auto / auto / span 2 / span 3"
           :video-name="internalVideoName"
-          :video-num="videoNum"
           @navigate="(event) => $emit('navigate', event)"
         />
         <div v-else-if="!loading" style="grid-area: search-results; line-height: 3; text-align: center; align-self: center">
